Implements new route for review controller.

+++ dealership-review-app/src/main/java/com/dealershipreviewapp/dealershipreviewapp/controllers/ReviewController.java
@@ -3,8 +3,13 @@ package com.dealershipreviewapp.dealershipreviewapp.controllers;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
 import com.dealershipreviewapp.dealershipreviewapp.entity.Review;
+import com.dealershipreviewapp.dealershipreviewapp.entity.User;
+import com.dealershipreviewapp.dealershipreviewapp.security.SecurityConstants;
 import com.dealershipreviewapp.dealershipreviewapp.service.ReviewService;
+import com.dealershipreviewapp.dealershipreviewapp.service.UserService;
 
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
@@ -18,6 +23,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 
 @AllArgsConstructor
@@ -26,6 +32,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 public class ReviewController {
 
     private ReviewService reviewService;
+    private UserService userService;
+    
     @GetMapping("/all")
     public ResponseEntity<Set<Review>> getReviews() {
         Set<Review> allReviews = reviewService.getAllReviews();
@@ -37,7 +45,7 @@ public class ReviewController {
         return new ResponseEntity<>(reviewService.createReview(review, userId), HttpStatus.CREATED);
     }
 
-    @GetMapping("user/{userId}")
+    @GetMapping("user/{userId}/all")
     public ResponseEntity<Set<Review>> getReviewsFromUser(@PathVariable Long userId) {
         return new ResponseEntity<>(reviewService.getReviewsByUser(userId), HttpStatus.OK);
     }
@@ -48,7 +56,20 @@ public class ReviewController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteReview(@PathVariable Long id) {
+    public ResponseEntity<HttpStatus> deleteReview(@PathVariable Long id, @RequestHeader("Authorization") String bearerToken) {
+        String token = bearerToken.replace(SecurityConstants.BEARER, "");
+        String username = JWT.require(Algorithm.HMAC512(
+                SecurityConstants.SECRET))
+                .build()
+                .verify(token)
+                .getSubject();
+        User user = userService.getUser(username);
+        Review review = reviewService.getReview(id);
+
+        if(review.getUser() != user) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         reviewService.deleteReview(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
